backend/App/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.29
openai==1.18.0
python-dotenv==1.0.1
httpx==0.24.1